# lib/api-zod/src/generated/types/teamMemberUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Task Center API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TeamMemberUpdate {
  name?: string;
  role?: string;
  phone?: string;
  email?: string;
  avatarUrl?: string;
}